Silence bogus MSVC signage warning in qtcompat.h
It's warning about comparing a size_t with that C++-flavored INT_MAX
thing, the latter obviously always being positive, but it doesn't
realize that fact. Now it converts it to a size_t.

+++ src/libshared/util/qtcompat.h
@@ -89,7 +89,7 @@ using RetrieveDataMetaType = QVariant::Type;
 using sizetype = int;
 
 inline auto castSize(size_t size) {
-	Q_ASSERT(size < std::numeric_limits<int>().max());
+	Q_ASSERT(size < size_t(std::numeric_limits<int>().max()));
 	return int(size);
 }
